Styling tables in dashboard.jsx and company.jsx
Added style to the table in dashboard as well.

Changed the classnames on the table in company.jsx so they also fit to the table in the dashboard.jsx.

Added these classnames on the table in the dashboard.jsx as well so this table get the same design and layout as the on in company.jsx

This will also prevent us from having duplicated code in the styledcomponents.jsx