Adjusting the menu and cards
Adjusting the menu so the user-tab in the admin menu view look better. Made it on the same line as the rest of the menu with the same design in styledcomponents.

Made the width of the cards flexible. When one card contains longer lines of information, for example a longer email address, the card will adjust and get wider.