Dashboard-cards Phase Number
Added code in the Server folder, in the files: company.controller.jsx, company.model.jsx and company.service.jsx to add Phase as a attribute when adding a new company, and updating companies.

Have already added code in the DashboardList.jsx in the Client-folder, to display the phase on the cards.