Buttons and Lists.
Improved the main buttons, and the edit- and delete-buttons in the tables in admin>manage users, and admin>manage companies.

Also added color on every second line in the tables, and they'll now turn grey when hovered. Making it easier to see what line you are pointing on with your mouse.