Button design in style.less and added a Back-button
Moved the styling of all the buttons to style.less.
Also had to change the linking of the classnames in every file from className="" to className={''}.

In addition, I added a 'Back' button in the admin view when visiting the table of user, and when visiting the table of companies. The back button will take the user back to  the start-admin-view.